pyproject.toml
[tool.poetry]
name = "tbot-by-prod-system"
version = "0.1.0"
description = "Product system"
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
aiogram = "^3.13.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
